water-well-spatial-explorer-app/app/requirements.txt
dash
dash-leaflet
databricks-sdk
psycopg[binary]
psycopg-pool
pandas
plotly
python-dotenv